requirements.txt
Django==1.10
djangorestframework
logger==1.4
raven
psycopg2-binary
urllib3
whitenoise==3.2
dj-static==0.0.6
gunicorn==19.7.1
pip install django-cors-headers